V3/requirements.txt
RPi.GPIO
smbus2
PyYAML

